chore: タグ周りのディレクトリ構造の見直し
