+++ src/features/manage/common/funcs/index.ts
@@ -1,4 +1,3 @@
 export * from "./convertDifficultyToNumber";
 export * from "./convertJPToENWeekday";
 export * from "./convertEnToJPWeekday";
-export * from "./selectTagColor";
