+++ src/features/manage/new/components/NewTagDropdown.tsx
@@ -2,7 +2,7 @@ import type { FC } from "react";
 import type { UseFormRegister, UseFormWatch } from "react-hook-form";
 import { TagItem } from "../../common/components/TagItem";
 import type { TManageValidationSchema } from "../../common/libs/validation";
-import { useQueryTagList } from "../../tags/api/tag/useQueryTag";
+import { useQueryTagList } from "../../../tags/api/tag/useQueryTag";
 import { toSelectColorLabel } from "../../common/funcs/toSelectColorLabel";
 
 type Props = {
