+++ src/features/tags/components/TagsColorDropdown.tsx
@@ -1,8 +1,8 @@
 import type { FC } from "react";
 import type { UseFormRegister, UseFormWatch } from "react-hook-form";
-import { selectTagColorText } from "../../common/funcs";
 import type { TTagValidationSchema } from "../libs/validation";
 import { TagsColorItem } from "./TagsColorItem";
+import { selectTagColorText } from "../funcs/selectTagColor";
 
 type Props = {
   register: UseFormRegister<TTagValidationSchema>;
